% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendTopics.R
\name{dendTopics}
\alias{dendTopics}
\title{Topic Dendrogram}
\usage{
dendTopics(sims, ind, method = "complete")
}
\arguments{
\item{sims}{[\code{symmetrical named matrix}]\cr
Pairwise jaccard similarities of underlying topics like
the output from \code{\link{jaccardTopics}}. The topic names should be
formatted as <\emph{Run X}>.<\emph{Topic Y}>, so that the name before the
first dot identifies the LDA run.}

\item{ind}{[\code{integer} or \code{character}]\cr
An integerish vector for specifying the topics taken into account. Alternatively
a character vector can be passed. Then, all topics are taken for which the name
contain at least one of the phrases in \code{ind} (see \code{\link[base]{grepl}}).}

\item{method}{[\code{character(1)}]\cr
The agglomeration method. See \code{\link[stats]{hclust}}.}
}
\value{
[\code{\link[stats]{dendrogram}}] of all considered topics.
}
\description{
Builds a dendrogram for topics based on their pairwise similarities using a
cluster algorithm.
}
\details{

}
\examples{
# TODO

}
