% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrototype.R
\name{getPrototype}
\alias{getPrototype}
\title{...}
\usage{
getPrototype(...)
}
\arguments{
\item{...}{additional arguments}

\item{x}{[\code{named list}]\cr
\code{\link{LDABatch}} or \code{\link{LDARep}} object.}

\item{lda}{[\code{named list}]\cr
List of \code{\link{LDA}} objects, named by the corresponding "job.id".}

\item{vocab}{.}

\item{limit.rel}{.}

\item{limit.abs}{.}

\item{progress}{.}

\item{keepTopics}{.}

\item{keepSims}{.}

\item{sclop}{.}
}
\value{
[\code{named list}] with entries ...
}
\description{
...
}
