% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSCLOP.R
\name{getSCLOP}
\alias{getSCLOP}
\alias{getSimilarity}
\alias{getMergedTopics}
\alias{getPrototypeID}
\alias{getLDA.PrototypeLDA}
\alias{getID.PrototypeLDA}
\alias{getParam.PrototypeLDA}
\title{Getter for PrototypeLDA}
\usage{
getSCLOP(x)

getSimilarity(x)

getMergedTopics(x)

getPrototypeID(x)

\method{getLDA}{PrototypeLDA}(x, job, reduce = TRUE)

\method{getID}{PrototypeLDA}(x)

\method{getParam}{PrototypeLDA}(x)
}
\arguments{
\item{x}{[\code{named list}]\cr
\code{\link[=getPrototype]{PrototypeLDA}} object.}

\item{job}{[\code{\link{data.frame}} or \code{integer}]\cr
A data.frame or data.table with a column named "job.id" or a vector of
integerish job ids. Default is the (integerish) ID of the Prototype LDA.}

\item{reduce}{[\code{logical(1)}]\cr
If the list of LDAs contains only one element, should the list be reduced and
the single (unnamed) element be returned? Default is \code{TRUE}.}
}
\description{
Returns the corresponding element of a \code{\link[=getPrototype]{PrototypeLDA}} object.
}
