% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimilarity.R
\name{getSimilarity}
\alias{getSimilarity}
\alias{getRelevantWords}
\alias{getConsideredWords}
\alias{getParam.TopicSimilarity}
\title{Getter for TopicSimilarity}
\usage{
getSimilarity(x)

getRelevantWords(x)

getConsideredWords(x)

\method{getParam}{TopicSimilarity}(x)
}
\arguments{
\item{x}{[\code{named list}]\cr
\code{\link[=jaccardTopics]{TopicSimilarity}} object.}
}
\description{
Returns the corresponding element of a \code{\link[=jaccardTopics]{TopicSimilarity}} object.
}
\seealso{
Other getter functions: \code{\link{getJob}},
  \code{\link{getSCLOP}}, \code{\link{getTopics}}
}
\concept{TopicSimilarity funtions}
\concept{getter functions}
