% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R, R/LDABatch.R, R/LDARep.R
\name{is.LDA}
\alias{is.LDA}
\alias{is.LDABatch}
\alias{is.LDARep}
\title{Checking Objects}
\usage{
is.LDA(x, verbose = FALSE)

is.LDABatch(x)

is.LDARep(x)
}
\description{
Checking Objects
}
