% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA}
\alias{LDA}
\title{LDA Object}
\usage{
LDA(x, assignments = NULL, topics = NULL, document_sums = NULL,
  document_expects = NULL, log.likelihoods = NULL)
}
\arguments{
\item{x}{[\code{named list}]\cr
Output from \code{\link[lda]{lda.collapsed.gibbs.sampler}}. Alternatively each
element can be passed for individual results. Individually set parameters
overwrite elements from \code{x}.}

\item{assignments}{.}

\item{topics}{.}

\item{document_sums}{.}

\item{document_expects}{.}

\item{log.likelihoods}{.}
}
\value{
[\code{named list}] as LDA object.
}
\description{
Constructor for LDA objects used in this package.
}
\details{

}
\examples{
#TODO

}
