% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTopics.LDARep.R
\name{mergeTopics.LDARep}
\alias{mergeTopics.LDARep}
\title{Merge LDA Topic Matrices}
\usage{
\method{mergeTopics}{LDARep}(x, vocab)
}
\arguments{
\item{x}{[\code{named list}]\cr
Output from \code{\link{LDARep}}.}

\item{vocab}{[\code{character}]\cr
Vocabularies taken into consideration for merging topic matrices.}
}
\value{
[\code{named matrix}] with the count of vocabularies (row wise) in topics (column wise).
}
\description{
Collects LDA results from a list of replicated runs and merges their topic
matrices for a given set of vocabularies.
}
\details{

}
\examples{
#TODO

}
