% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTopics.R
\name{mergeTopics}
\alias{mergeTopics}
\title{Merge LDA Topic Matrices}
\usage{
mergeTopics(...)
}
\arguments{
\item{...}{Arguments \code{x}, \code{vocab} and \code{progress} for
\code{\link{mergeRepTopics}} or \code{\link{mergeBatchTopics}}.}
}
\value{
[\code{named matrix}] with the count of vocabularies (row wise) in topics (column wise).
}
\description{
Generic function, which collects LDA results and merges their topic matrices for a
given set of vocabularies. Uses the function \code{\link{mergeRepTopics}}
or \code{\link{mergeBatchTopics}}.
}
\details{

}
\examples{
#TODO

}
