% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMS.R
\name{MMS}
\alias{MMS}
\alias{pruneMMS}
\alias{disparitySum}
\alias{MMS.pairwise}
\title{Mixed Memberships Similarity/Stability}
\usage{
MMS(dend)

pruneMMS(dend)

disparitySum(dend)

MMS.pairwise(sims)
}
\arguments{
\item{dend}{[\code{\link[stats]{dendrogram}}]\cr
Output from \code{\link{dendTopics}}.}

\item{sims}{[\code{symmetrical named matrix}]\cr
Pairwise jaccard similarities of underlying topics like
the output from \code{\link{jaccardTopics}}. The topic names should be
formatted as <\emph{Run X}>.<\emph{Topic Y}>, so that the name before the
first dot identifies the LDA run.}
}
\value{
\describe{
  \item{\code{MMS}}{[0,1] value specifying the MMS for the best possible
  local pruning state of the given dendrogram.}
  \item{\code{pruneMMS}}{[\code{list of \link[stats]{dendrogram}s}] }
  \item{\code{disparitySum}}{[\code{numeric(1)}] value specifying the least
  possible sum of disparities on the given dendrogram.}
  \item{\code{MMS.pairwise}}{[\code{symmetrical named matrix}] with all
  pairwise MMS scores of the given LDA runs.}
}
}
\description{
The function \code{MMS} calculates the MMS for the best possible local pruning state
of a dendrogram from \code{\link{dendTopics}}. The function \code{pruneMMS}
supplies the corresponding pruning state itself.\cr
To get all pairwise MMS scores of two LDA runs, the function \code{MMS.pairwise}
can be used. It returnes a matrix of the pairwise MMS scores.\cr
All three functions use the function \code{disparitySum} to calculate the
least possible sum of disparities (on the best possible local pruning state)
on a given dendrogram.
}
\details{

}
\examples{
# TODO

}
