% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCLOP.R
\name{SCLOP}
\alias{SCLOP}
\alias{disparitySum}
\alias{SCLOP.pairwise}
\title{Similarity/Stability of multiple sets of Objects using Clustering with Local Pruning}
\usage{
SCLOP(dend)

disparitySum(dend)

SCLOP.pairwise(sims)
}
\arguments{
\item{dend}{[\code{\link[stats]{dendrogram}}]\cr
Output from \code{\link{dendTopics}}.}

\item{sims}{[\code{lower triangular named matrix}]\cr
Pairwise jaccard similarities of underlying topics as the \code{sims} element
from \code{\link[=jaccardTopics]{TopicSimilarity}} objects. The topic names should be
formatted as <\emph{Run X}>.<\emph{Topic Y}>, so that the name before the
first dot identifies the LDA run.}
}
\value{
\describe{
  \item{\code{SCLOP}}{[0,1] value specifying the S-CLOP for the best possible
  local pruning state of the given dendrogram.}
  \item{\code{disparitySum}}{[\code{numeric(1)}] value specifying the least
  possible sum of disparities on the given dendrogram.}
  \item{\code{SCLOP.pairwise}}{[\code{symmetrical named matrix}] with all
  pairwise S-CLOP scores of the given LDA runs.}
}
}
\description{
The function \code{SCLOP} calculates the S-CLOP for the best possible local pruning state
of a dendrogram from \code{\link{dendTopics}}. The function \code{\link{pruneSCLOP}}
supplies the corresponding pruning state itself.\cr
To get all pairwise S-CLOP scores of two LDA runs, the function \code{SCLOP.pairwise}
can be used. It returnes a matrix of the pairwise S-CLOP scores.\cr
All three functions use the function \code{disparitySum} to calculate the
least possible sum of disparities (on the best possible local pruning state)
on a given dendrogram.
}
\details{

}
\examples{
# TODO

}
