% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.LDARep.R
\name{as.LDARep}
\alias{as.LDARep}
\alias{as.LDARep.default}
\alias{as.LDARep.LDABatch}
\alias{is.LDARep}
\title{LDARep Constructor}
\usage{
as.LDARep(x, lda, job, id)

\method{as.LDARep}{default}(lda, job, id)

\method{as.LDARep}{LDABatch}(x)

is.LDARep(obj, verbose = FALSE)
}
\arguments{
\item{x}{[\code{named list}]\cr
\code{\link{LDABatch}} or \code{\link{LDARep}} object.}

\item{lda}{[\code{named list}]\cr
List of \code{\link{LDA}} objects, named by the corresponding "job.id" (\code{integerish}).}

\item{job}{[\code{\link{data.frame}} or \code{named vector}]\cr
A data.frame or data.table with named columns "job.id" (\code{integerish}), "K", "alpha", "eta"
and "num.iterations" or a named vector with entries "K", "alpha", "eta" and
"num.iterations".}

\item{id}{[\code{character(1)}]\cr
A name for the computation. If not passed, is set to "LDARep".}
}
\value{
[\code{named list}] with entries \code{id} for computation's name,
\code{jobs} for the parameter settings and \code{lda} for the results themselves.
}
\description{
Constructs a \code{\link{LDARep}} object for given elements \code{lda},
\code{job} and \code{id}.
}
\details{

}
\examples{
#TODO

}
