% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.LDARep.R
\name{as.LDARep}
\alias{as.LDARep}
\alias{is.LDARep}
\title{LDARep Constructor}
\usage{
as.LDARep()

is.LDARep(obj, verbose = FALSE)
}
\value{
[\code{named list}] with entries
}
\description{
Constructs a \code{\link{LDARep}} object for given elements ...
}
\details{

}
\examples{
#TODO

}
