% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendTopics.R
\name{dendTopics}
\alias{dendTopics}
\alias{plot.TopicDendrogram}
\title{Topic Dendrogram}
\usage{
dendTopics(sims, ind, method = "complete")

\method{plot}{TopicDendrogram}(x, pruning, pruning.par, ...)
}
\arguments{
\item{sims}{[\code{\link[=jaccardTopics]{TopicSimilarity}} object
or \code{lower triangular named matrix}]\cr
\code{\link[=jaccardTopics]{TopicSimilarity}} object or
pairwise jaccard similarities of underlying topics as the \code{sims} element
from \code{\link[=jaccardTopics]{TopicSimilarity}} objects. The topic names should be
formatted as <\emph{Run X}>.<\emph{Topic Y}>, so that the name before the
first dot identifies the LDA run.}

\item{ind}{[\code{integer} or \code{character}]\cr
An integerish vector for specifying the topics taken into account. Alternatively
a character vector can be passed. Then, all topics are taken for which the name
contain at least one of the phrases in \code{ind} (see \code{\link[=grep]{grepl}}).
By default all topics are considered.}

\item{method}{[\code{character(1)}]\cr
The agglomeration method. See \code{\link[stats]{hclust}}.}

\item{x}{an R object.}

\item{pruning}{[\code{list of \link[stats]{dendrogram}s}]\cr
\code{\link[=pruneSCLOP]{PruningSCLOP}} object specifying the best possible
local pruning state.}

\item{pruning.par}{[\code{list}]\cr
List of parameters to mark the pruning. See \code{\link[=pruneSCLOP]{pruning.par}}
for default parameters. Types for marking the pruning state are \code{"abline"},
\code{"color"} and \code{"both"}.}

\item{...}{additional arguments.}
}
\value{
[\code{\link[stats]{dendrogram}}] \code{\link[=dendTopics]{TopicDendrogram}}
(and \code{\link[stats]{dendrogram}}) object of all considered topics.
}
\description{
Builds a dendrogram for topics based on their pairwise similarities using a
cluster algorithm.
}
\details{
The label´s colors are determined based on their Run belonging using
\code{\link[colorspace]{rainbow_hcl}} by default. Colors can be manipulated
using \code{\link[dendextend]{labels_colors}}. Analogously, the labels
themself can be  manipulated using \code{\link[dendextend:labels.hclust]{labels}}.
For both the function \code{\link[stats]{order.dendrogram}} is useful.
}
\examples{
# TODO

}
\seealso{
Other plot functions: \code{\link{pruneSCLOP}}
}
\concept{plot functions}
