% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJob.R
\name{getJob}
\alias{getJob}
\alias{getID}
\alias{getRegistry}
\alias{setFileDir}
\title{Get Registry and Jobs}
\usage{
getJob(x)

getID(x)

getRegistry(x)

setFileDir(x, file.dir)
}
\arguments{
\item{x}{[\code{named list}]\cr
Output from \code{\link{LDABatch}}.}

\item{file.dir}{[\code{character(1)}]\cr
New file directory to overwrite the registry's old one. This can be useful
if the registry is transferred from a batch system.}
}
\description{
Returns the job ids and its parameter set (\code{getJob}) or the registry's
id (\code{getID}) or the registry itself (\code{getRegistry}) for a
\code{\link{LDABatch}} object.
}
