% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJob.R
\name{getJob}
\alias{getJob}
\title{Get Registry and Jobs}
\usage{
getJob(x)
}
\description{
Returns the job ids and its parameter set (\code{getJob}) or the registry's
id (\code{getID}) or the registry itself (\code{getRegistry}).
}
