% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSCLOP.R
\name{getSCLOP}
\alias{getSCLOP}
\alias{getSimilarity}
\alias{getMergedTopics}
\alias{getLDAID}
\alias{getLDA.PrototypeLDA}
\alias{getID.PrototypeLDA}
\alias{getparam.PrototypeLDA}
\title{Getter for PrototypeLDA}
\usage{
getSCLOP(x)

getSimilarity(x)

getMergedTopics(x)

getLDAID(x)

\method{getLDA}{PrototypeLDA}(x, job, reduce)

\method{getID}{PrototypeLDA}(x)

getparam.PrototypeLDA(x)
}
\arguments{
\item{x}{[\code{named list}]\cr
\code{\link[getPrototype]{PrototypeLDA}} object.}

\item{job}{not implemented (relict from \code{\link{getLDA}})}

\item{reduce}{not implemented (relict from \code{\link{getLDA}})}
}
\description{
Returns the corresponding element of a \code{\link[getPrototype]{PrototypeLDA}} object.
}
