% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopics.R
\name{getTopics}
\alias{getTopics}
\alias{getAssignments}
\alias{getDocument_sums}
\alias{getDocument_expects}
\alias{getLog.likelihoods}
\alias{getAlpha}
\alias{getEta}
\alias{getK}
\alias{getNum.iterations}
\alias{getParam}
\alias{getEstimators}
\title{Getter for LDA}
\usage{
getTopics(x)

getAssignments(x)

getDocument_sums(x)

getDocument_expects(x)

getLog.likelihoods(x)

getAlpha(x)

getEta(x)

getK(x)

getNum.iterations(x)

getParam(x)

getEstimators(x)
}
\arguments{
\item{x}{[\code{named list}]\cr
\code{\link{LDA}} object.}
}
\description{
Returns the corresponding element of a \code{\link{LDA}} object.
}
