% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopics.R
\name{getTopics}
\alias{getTopics}
\alias{getAssignments}
\alias{getDocument_sums}
\alias{getDocument_expects}
\alias{getLog_likelihoods}
\title{Getter for LDA}
\usage{
getTopics(x)

getAssignments(x)

getDocument_sums(x)

getDocument_expects(x)

getLog_likelihoods(x)
}
\arguments{
\item{x}{[\code{named list}]\cr
\code{\link{LDA}} object.}
}
\description{
Returns the corresponding element of a \code{\link{LDA}} object.
}
