% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardTopics.R
\name{jaccardTopics}
\alias{jaccardTopics}
\title{Pairwise Jaccard Coefficients}
\usage{
jaccardTopics(topics, limit.rel = 1/500, limit.abs = 10,
  progress = TRUE)
}
\arguments{
\item{topics}{[\code{named matrix}]\cr
The counts of vocabularies (row wise) in topics (column wise).}

\item{limit.rel}{[0,1]\cr
A relative lower bound limit for which words are taken into account. All words
are taken as relevant for a topic that have a count higher than \code{limit.rel}
multiplied by the total count of the given topic.}

\item{limit.abs}{[\code{integer(1)}]\cr
An absolute lower bound limit for which words are taken into account. All words
are taken as relevant for a topic that have a count higher than \code{limit.abs}.}

\item{progress}{[\code{logical(1)}]\cr
Should a nice progress bar be shown? Turning it off, could lead to significantly
faster calculation. Default ist \code{TRUE}.}
}
\value{
[\code{lower triangular named matrix}] with all pairwise jaccard similarities
of the given topics.
}
\description{
Calculates the similarity of all pairwise topic combinations using a modified
Jaccard Coefficient.
}
\details{

}
\examples{
# TODO

}
