% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA}
\alias{LDA}
\alias{as.LDA}
\alias{is.LDA}
\title{LDA Object}
\usage{
LDA(x, param, assignments, topics, document_sums, document_expects,
  log.likelihoods)

as.LDA(x, param, assignments, topics, document_sums, document_expects,
  log.likelihoods)

is.LDA(obj, verbose = FALSE)
}
\arguments{
\item{x}{[\code{named list}]\cr
Output from \code{\link[lda]{lda.collapsed.gibbs.sampler}}. Alternatively each
element can be passed for individual results. Individually set elements
overwrite elements from \code{x}.}

\item{param}{[\code{named list}]\cr
Parameters of the function call \code{\link[lda]{lda.collapsed.gibbs.sampler}}.
List always should contain names "K", "alpha", "eta" and "num.iterations".}

\item{assignments}{Individual element for LDA object.}

\item{topics}{Individual element for LDA object.}

\item{document_sums}{Individual element for LDA object.}

\item{document_expects}{Individual element for LDA object.}

\item{log.likelihoods}{Individual element for LDA object.}

\item{obj}{[\code{R} object]\cr
Object to test.}

\item{verbose}{[\code{logical(1)}]\cr
Should test information be given in the console?}
}
\value{
[\code{named list}] LDA object.
}
\description{
Constructor for LDA objects used in this package.
}
\details{

}
\examples{
#TODO

}
\seealso{
Other constructor functions: \code{\link{as.LDABatch}},
  \code{\link{as.LDARep}}

Other LDA functions: \code{\link{LDABatch}},
  \code{\link{LDARep}}, \code{\link{getTopics}}
}
\concept{LDA functions}
\concept{constructor functions}
