% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA}
\alias{LDA}
\alias{as.LDA}
\alias{is.LDA}
\title{LDA Object}
\usage{
LDA(x, assignments = NULL, topics = NULL, document_sums = NULL,
  document_expects = NULL, log.likelihoods = NULL)

as.LDA(x, assignments = NULL, topics = NULL, document_sums = NULL,
  document_expects = NULL, log.likelihoods = NULL)

is.LDA(obj, verbose = FALSE)
}
\arguments{
\item{x}{[\code{named list}]\cr
Output from \code{\link[lda]{lda.collapsed.gibbs.sampler}}. Alternatively each
element can be passed for individual results. Individually set elements
overwrite elements from \code{x}.}

\item{assignments}{Individual element for LDA object.}

\item{topics}{Individual element for LDA object.}

\item{document_sums}{Individual element for LDA object.}

\item{document_expects}{Individual element for LDA object.}

\item{log.likelihoods}{Individual element for LDA object.}

\item{obj}{[\code{R} object]\cr
Object to test.}

\item{verbose}{[\code{integer(1)}]\cr
Should test information be given in the console?}
}
\value{
[\code{named list}] LDA object.
}
\description{
Constructor for LDA objects used in this package.
}
\details{

}
\examples{
#TODO

}
