% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDABatch.R
\name{LDABatch}
\alias{LDABatch}
\title{LDA Replications on a Batch System}
\usage{
LDABatch(id = "LDABatch", docs, vocab, n = 100, seeds, load = FALSE,
  chunk.size = 1, resources, ...)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
Name for the registry's folder.}

\item{docs}{[\code{list}]\cr
Documents as received from \code{\link[tosca]{LDAprep}}.}

\item{vocab}{[\code{character}]\cr
Vocabularies passed to \code{\link[lda]{lda.collapsed.gibbs.sampler}}.}

\item{n}{[\code{integer(1)}]\cr
Number of Replications.}

\item{seeds}{[\code{integer(n)}]\cr
Random Seeds for each Replication.}

\item{load}{[\code{logical(1)}]\cr
If a folder with name \code{id} exists: should the existing registry be loaded?}

\item{chunk.size}{[\code{integer(1)}]\cr
Requested number of chunks. See \code{\link[batchtools]{chunk}}.}

\item{resources}{[\code{named list}]\cr
Computational resources for the jobs to submit. See \code{\link[batchtools]{submitJobs}}.}

\item{...}{additional arguments passed to \code{\link[lda]{lda.collapsed.gibbs.sampler}}.
Arguments may be passed as scalar (which will be coerced to a vector of length \code{n})
or vector of length \code{n}.}
}
\value{
[\code{named list}] with entries \code{id} for the registry's folder name,
\code{ids} for the submitted jobs' ids and \code{reg} for the registry itself.
}
\description{
Performs multiple runs of Latent Dirichlet Allocation on a batch system using
the package \code{\link[batchtools]{batchtools}}.
}
\details{

}
\examples{
#TODO

}
