% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTopics.R
\name{mergeTopics}
\alias{mergeTopics}
\title{Merge LDA Topic Matrices}
\usage{
mergeTopics(x)
}
\arguments{
\item{x}{[\code{named list}]\cr
Output from \code{\link{LDARep}} or \code{\link{LDABatch}}.}
}
\value{
[\code{named matrix}] with the count of vocabularies (row wise) in topics (column wise).
}
\description{
Collects LDA results and merges their topic matrices for a given set of vocabularies.
}
\details{

}
\examples{
#TODO

}
