% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reuters}
\alias{reuters}
\alias{reuters_docs}
\alias{reuters_vocab}
\alias{docs}
\alias{vocab}
\title{A Snippet of the Reuters Dataset}
\format{An object of class \code{list} of length 99.}
\source{
\url{http://ronaldo.cs.tcd.ie/esslli07/data/reuters21578-xml/}
}
\usage{
data(reuters_docs)

data(reuters_vocab)
}
\description{
bla
}
\references{
Lewis, David (1997). \emph{Reuters-21578 Text Categorization Collection Distribution 1.0}.
\url{http://kdd.ics.uci.edu/databases/reuters21578/reuters21578.html}
Luz, Saturnino. \emph{XML-encoded version of Reuters-21578}.
\url{http://ronaldo.cs.tcd.ie/esslli07/data/reuters21578-xml/}
}
\keyword{datasets}
